changed name of home component to AllProducts & added Home component
